+++ R/messages.R
@@ -260,7 +260,19 @@ messages$warnParameterMetadata <- function(message) {
   )
 }
 
-messages$messageOptionsReset <- function(field, oldValue, newValue, optionsField) {
+messages$errorParameterMetadataMissing <- function() {
+  paste0(
+    "Parameter metadata is not available. `toDataFrame()` cannot be called ",
+    "on a `PIResult` created without `piParameters`."
+  )
+}
+
+messages$messageOptionsReset <- function(
+  field,
+  oldValue,
+  newValue,
+  optionsField
+) {
   ospsuite.utils::cliFormat(
     "{.field {field}} changed from {.val {oldValue}} to {.val {newValue}}. {.field {optionsField}} reset to defaults."
   )
@@ -269,5 +281,11 @@ messages$messageOptionsReset <- function(field, oldValue, newValue, optionsField
 messages$warningUnknownOptions <- function(keys, fieldName) {
   ospsuite.utils::cliFormat(
     "{cli::qty(length(keys))}Unknown option{?s} for {.field {fieldName}}: {.val {keys}}. Ignored."
+  )
+}
+
+messages$warningNoValidErrorValues <- function() {
+  ospsuite.utils::cliFormat(
+    "Some error values are invalid and will use unit weights (equivalent to {.val none})."
   )
 }